Miscellaneous fixes (#4584)
* Fix use after move in LevelDbPersistenceForTesting

* Fix size-related errors when compiling in 32-bit mode

* Fix test-only memory leaks

* Fix test-only memory leaks

* Prevent spec tests from appearing to leak

The spec tests aren't actually leaking, but they accumulate enough
garbage without clearing the autorelease pool that Instruments
complains.

* Avoid temporaries in methods that @throw

ARC does not release temporaries when an Objective-C exception is
thrown. Use C++ equivalents directly to avoid leaks as detected by
Instruments.

* Fix memory leak in acknowledge mutation

Any previous value must be freed before the next one is assigned.

* Manually implement assignment operators in ByteString

The single by-value implementation wasn't preventing the compiler
generated standard versions from being generated which was causing
leaks.

* Prevent leaking GrpcCompletions that complete after Shutdown

After shutdown, the AsyncQueue silently drops Enqueue requests which
prevents these blocks from being deleted.

* Avoid racing with enqueued operations

It's possible for the Executor to immediately execute the given block,
at which point the block might observe the async queue in a
not-shutting-down state.

+++ Firestore/core/test/firebase/firestore/local/lru_garbage_collector_test.cc
@@ -655,7 +655,7 @@ TEST_P(LruGarbageCollectorTest, RemoveTargetsThenGC) {
 TEST_P(LruGarbageCollectorTest, GetsSize) {
   NewTestResources();
 
-  size_t initial_size = gc_->CalculateByteSize();
+  int64_t initial_size = gc_->CalculateByteSize();
 
   persistence_->Run("fill cache", [&] {
     // Simulate a bunch of ack'd mutations.
@@ -665,7 +665,7 @@ TEST_P(LruGarbageCollectorTest, GetsSize) {
     }
   });
 
-  size_t final_size = gc_->CalculateByteSize();
+  int64_t final_size = gc_->CalculateByteSize();
   ASSERT_GT(final_size, initial_size);
 }
